Move terminalQuickFixMenuItems into quickFixAddon
